Use UK geography resolver for scoped regions
