+++ tests/test_local_authority_impact.py
@@ -178,7 +178,7 @@ def test_compute_local_authority_impacts_local_only_does_not_call_gcs(tmp_path):
 
     with (
         patch(
-            "policyengine.outputs.uk_geography_assets.default_local_search_dirs",
+            "policyengine.data.uk_geography_assets.default_local_search_dirs",
             return_value=[tmp_path / "missing"],
         ),
         patch("policyengine_core.tools.google_cloud.download_gcs_file") as download,
